+++ components/SelectionColorOverride.tsx
@@ -3,7 +3,7 @@ import * as React from "react"
 export default function SelectionColorOverride() {
   return <style dangerouslySetInnerHTML={{
     __html: [
-      "::selection {",
+      "svg text::selection {",
       "color: transparent;",
       "background-color: rgba(130, 130, 130, .5);",
       "}",