+++ components/SvgAnalysis.tsx
@@ -1,11 +1,8 @@
-import {effectiveTextOrientation, scaleRectangle} from "../util/OverlayUtil"
 import * as React from "react"
-import {useContext} from "react"
 import {JpdbCardState} from "../model/JpdbCardState"
 import {ImageAnalysisParagraph} from "../model/ImageAnalysis"
 import {getJpdbVocabularyCardStates, JpdbVocabulary} from "../model/JpdbVocabulary"
-import {SvgOverlayContext} from "../util/SvgOverlayContext"
-import SvgSymbol from "./SvgSymbol"
+import SvgAnalysisFragment from "./SvgAnalysisFragment"
 
 interface Props {
   paragraphs: readonly ImageAnalysisParagraph[],
@@ -14,35 +11,22 @@ interface Props {
 }
 
 export default function SvgAnalysis({paragraphs, vocabulary, showAnalysis}: Props) {
-  const {scaleX, scaleY, textOrientation} = useContext(SvgOverlayContext)
-
   return <>{
     paragraphs.flatMap(paragraph =>
-      paragraph.fragments.flatMap((fragment, fragmentIndex) => {
-        const bounds = scaleRectangle(fragment.bounds, scaleX, scaleY)
-        const color = getColorForState(getJpdbVocabularyCardStates(vocabulary, fragment.vocabularyIndex))
-        return <g key={`ag-${paragraph.id}-${fragmentIndex}`}>
-          {color && showAnalysis ? <rect
-            x={bounds.x}
-            y={bounds.y}
-            width={bounds.w}
-            height={bounds.h}
-            style={{fill: color}}
-          /> : null}
-          {fragment.symbols.map((symbol, symbolIndex) =>
-            <SvgSymbol
-              key={`as-${paragraph.id}-${fragmentIndex}-${symbolIndex}`}
-              packedSymbol={symbol}
-              textOrientation={effectiveTextOrientation(textOrientation, fragment.orientation)}
-            />
-          )}
-        </g>
-      })
+      paragraph.fragments.flatMap((fragment, fragmentIndex) =>
+        <SvgAnalysisFragment
+          key={`ag-${paragraph.id}-${fragmentIndex}`}
+          fragment={fragment}
+          color={getColorForState(getJpdbVocabularyCardStates(vocabulary, fragment.vocabularyIndex))}
+          showAnalysis={showAnalysis}
+          vocabulary={vocabulary[fragment.vocabularyIndex]}
+        />
+      )
     )
   }</>
 }
 
-function getColorForState(states: JpdbCardState[]) {
+function getColorForState(states: JpdbCardState[]): string | undefined {
   switch (states[0]) {
     case JpdbCardState.Learning:
       return "rgba(74,231,129,0.15)"
@@ -52,6 +36,6 @@ function getColorForState(states: JpdbCardState[]) {
     case JpdbCardState.New:
       return "rgba(32,48,145,0.15)"
     default:
-      return null
+      return undefined
   }
 }