+++ components/SvgHighlight.tsx
@@ -1,11 +1,13 @@
 import * as React from "react"
-import {useState} from "react"
+import {useContext, useState} from "react"
 import {JpdbVocabulary} from "../model/JpdbVocabulary"
 import {Rectangle} from "../model/Rectangle"
-import {Portal, usePopper,} from "@chakra-ui/react"
+import {Portal, usePopper} from "@chakra-ui/react"
 import {JpdbPopupDialog} from "./JpdbPopup"
 import {JpdbRule} from "../model/JpdbRule"
 import {JpdbPopup} from "../model/JpdbPopup"
+import {SvgOverlayContext} from "../util/SvgOverlayContext"
+import {rgba} from "color2k"
 
 interface Props {
   bounds: Rectangle,
@@ -15,6 +17,8 @@ interface Props {
 }
 
 export default function SvgHighlight({bounds, rule, vocabulary, mouseOverGroup}: Props) {
+  const {jpdbMiningDeckId} = useContext(SvgOverlayContext)
+
   const {popperRef, referenceRef} = usePopper({placement: "right"})
 
   const [mouseOverPopup, setMouseOverPopup] = useState(false)
@@ -28,13 +32,14 @@ export default function SvgHighlight({bounds, rule, vocabulary, mouseOverGroup}:
       y={bounds.y}
       width={bounds.w}
       height={bounds.h}
-      style={{fill: rule.overlayColor}}
+      style={{fill: rule.overlayColor || rgba(0, 0, 0, 0)}}
       ref={referenceRef}
     />
     {popupOpen && popupPossible && <Portal>
       <JpdbPopupDialog
         ref={popperRef}
         vocabulary={vocabulary}
+        miningDeckId={jpdbMiningDeckId}
         compact={rule.popup === JpdbPopup.Compact}
         onMouseOver={(over) => setMouseOverPopup(over)}
       />