+++ CHANGES.md
@@ -3,7 +3,9 @@
   - If you are using this integration you will need to go the settings and enter your JPDB API key.
     SID cookie is no longer used.
   - The `data/.jpdb-cache` directory is no longer used and should be deleted. Cache is now in memory.
-  - Confidence selector now also works on analysis result
+  - Added JPDB vocabulary popup when hovering over the highlighted word.
+  - Confidence selector now also works on analysis result.
+- Text selection color override is now only applied to the SVG.
 
 #### Version 1.2.2